정렬이 제대로 되지않는 문제 수정
