+++ src/main/java/com/itac/login/control/ApiSearchController.java
@@ -3,6 +3,7 @@ package com.itac.login.control;
 import com.itac.login.entity.store.Store;
 import com.itac.login.service.StoreService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,10 +13,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
+import static com.itac.login.common.util.SortMethodClass.recommendfy;
+
 
 @RequestMapping("/api")
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class ApiSearchController {
 
     private final StoreService storeService;
@@ -59,4 +63,20 @@ public class ApiSearchController {
 //    }
 
     //평점이 좋은순
+
+
+    //추천 알고리즘이 정상적으로 동작하는지 확인하는 TestMapping
+    @GetMapping("/searchrecommend")
+    public ResponseEntity<Object> searchrecommend(){
+        List<Store> storeList = storeService.allStores();
+        log.info("before recommendfy storeList.toString() : "+storeList.toString());
+
+        storeList = recommendfy(storeList);
+        log.info("after recommendfy storeList.toString() : "+storeList.toString());
+
+        if(!storeList.isEmpty()){
+            return ResponseEntity.ok().body(storeList);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("조회된 데이터 없음");
+    }
 }
