+++ src/main/java/com/itac/login/service/StoreService.java
@@ -17,9 +17,10 @@ import java.util.List;
 public class StoreService {
     private final StoreRepository storeRepository;
     private Sort gradeDesc = Sort.by(Sort.Direction.DESC, "grade");
+    private Sort gradeAsc = Sort.by(Sort.Direction.ASC, "grade");
 
     public List<Store> allStores() {
-        return storeRepository.findAll(gradeDesc);
+        return storeRepository.findAll(gradeAsc);
     }
 
     public List<Store> searchWithWord(String searchWord) {
@@ -64,45 +65,5 @@ public class StoreService {
         return false;
     }
 
-    private List<Store> recommend(List<Store> list){
-        final int gradeWeight = 1;
-        HashMap<Store,Integer> map = new HashMap<Store,Integer>();
-        
-        //별점순으로 점수 매겨주기
-        list = sortByGrade(list);
-        for(Store store:list){
-            map.put(store,map.getOrDefault(store,0)+
-                    ((list.size()-1)-list.indexOf(store))/(list.size()-1)*100*gradeWeight);
-        }
 
-        //리뷰 많은순 처리 필요
-        //list =
-
-        //가격 낮은순으로 처리 필요
-
-        //거리 순으로 처리 필요
-
-        //처리된 결과에 따라 list 재정렬
-
-        Collections.sort(list,new Comparator<Store>(){
-            @Override
-            public int compare(Store o1, Store o2) {
-                return map.get(o2)-map.get(o1);
-            }
-        });
-
-        return list;
-    }
-
-    //점수 높은 순(내림차순)으로 정렬, 2번째 매개변수가 있을경우에는 (오름차순)으로 정렬
-    private List<Store> sortByGrade(List<Store> list,Object ...flag){
-
-        final int orderDirection = (flag.length==0?-1:1); //내림차순일 때 -1, 오름차순일때 1
-
-        list.sort((o1, o2) ->{
-            return orderDirection*(int)(Float.parseFloat(o1.getGrade())-Float.parseFloat(o2.getGrade()));
-        } );
-
-        return list;
-    }
 }
